requirements.txt
PySide6>=6.6
opencv-python>=4.9
numpy>=1.26
ultralytics>=8
psutil>=5.9
pyserial>=3.5
cv2-enumerate-cameras>=1.2.1
